make redhat conversion tool downloadable with Pip
