+++ tools/redhat/.gitignore
@@ -1,0 +1,1 @@
+redhat_osv.egg-info/
